+++ lib/maru/plugs/router.ex
@@ -5,11 +5,10 @@ defmodule Maru.Plugs.Router do
   def init(opts) do
     router = opts |> Keyword.fetch! :router
     'Elixir.' ++ _ = Atom.to_char_list router
-    version = opts |> Keyword.fetch!(:version)
+    version = opts |> Keyword.fetch! :version
     %Resource{path: path, param_context: param_context} = opts |> Keyword.get :resource, %Resource{}
     {router, path, version, param_context}
   end
-
 
   def call(%Conn{private: %{maru_version: v1}}=conn_orig, {router, path, v2, param_context})
   when is_nil(v2) or v1 == v2 do