+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ## v0.5.1-dev
+* Enhancements
+  * router extend support
 
 ## v0.5.0 (2015-7-23)
 * Enhancements